+++ modules/akai-fire/src/main/java/com/oikoaudio/fire/ViewCursorControl.java
@@ -20,6 +20,7 @@ public class ViewCursorControl {
     private final PinnableCursorClip selectedClip;
     private final DeviceBank drumBank;
     private final DrumPadBank drumPadBank;
+    private final DrumPadBank drumMeterPadBank;
     private final TrackBank trackBank;
     private int selectedTrackIndex = -1;
     private int selectedClipSlotIndex = -1;
@@ -100,6 +101,10 @@ public class ViewCursorControl {
         drumBank.setDeviceMatcher(drumMatcher);
         // drumDevice = drumBank.getItemAt(0);
         drumPadBank = primaryDevice.createDrumPadBank(16);
+        drumMeterPadBank = primaryDevice.createDrumPadBank(16);
+        drumPadBank
+                .scrollPosition()
+                .addValueObserver(position -> drumMeterPadBank.scrollPosition().set(position));
     }
 
     public TrackBank getTrackBank() {
@@ -147,5 +152,9 @@ public class ViewCursorControl {
 
     public DrumPadBank getDrumPadBank() {
         return drumPadBank;
+    }
+
+    public DrumPadBank getDrumMeterPadBank() {
+        return drumMeterPadBank;
     }
 }